Move pif job
