+++ @xen-orchestra/lite/src/modules/pif/jobs/pif-args.ts
@@ -1,0 +1,7 @@
+import type { XenApiPif } from '@/libs/xen-api/xen-api.types.ts'
+import { defineJobArg } from '@core/packages/job'
+
+export const pifsArg = defineJobArg({
+  identify: (pif: XenApiPif) => pif.$ref,
+  toArray: true,
+})
